Big update,logic in GameManager,Particles,Structure
Moved a lot of logic to GameManager which is now created by the
Game-class. This will make it easier to then move it right to the server.

Added some basic particle effects cause I wanted stuff to happen on
screen.

Started moving classes in to different maps to organize some structure,
could be more and improved upon.